Changed home screen
Added a two column screen with a sticky navigation bar. The left column has a gradient that gets darker as you scroll. The right column has two halftone SVGs: a base SVG with a light grey color, and another SVG with a dark grey color that is masked to the position of your mouse.